+++ cli/src/installers/index.ts
@@ -1,3 +1,5 @@
+import { payloadCMSInstaller } from "@/installers/payloadcms.js"
+import { typescriptInstaller } from "@/installers/typescript.js"
 import { type PackageManager } from "@/utils/get-user-pkg-manager.js"
 
 import { envVariablesInstaller } from "./env-vars.js"
@@ -14,12 +16,16 @@ export const availablePackages = [
   "envVariables",
   "prettier",
   "eslint",
+  "typescript",
+  "payload",
 ] as const
 export type AvailablePackages = (typeof availablePackages)[number]
 
+export const backendFrameworks = ["nextjs", "payload"] as const
+export type BackendFramework = (typeof backendFrameworks)[number]
+
 export const databaseProviders = ["sqlite", "mysql", "postgresql"] as const
 export type DatabaseProvider = (typeof databaseProviders)[number]
-
 export interface InstallerOptions {
   projectDir: string
   projectName: string
@@ -61,5 +67,13 @@ export const buildPkgInstallerMap = (
   eslint: {
     inUse: true,
     installer: eslintInstaller,
+  },
+  typescript: {
+    inUse: true,
+    installer: typescriptInstaller,
+  },
+  payload: {
+    inUse: packages.includes("payload"),
+    installer: payloadCMSInstaller,
   },
 })