+++ cli/template/packages/config/payload/with-sqlite.ts
@@ -1,0 +1,42 @@
+// storage-adapter-import-placeholder
+import path from "path"
+import { fileURLToPath } from "url"
+import { sqliteAdapter } from "@payloadcms/db-sqlite"
+import { payloadCloudPlugin } from "@payloadcms/payload-cloud"
+import { lexicalEditor } from "@payloadcms/richtext-lexical"
+import { buildConfig } from "payload"
+import sharp from "sharp"
+
+import { Media } from "@/collections/Media"
+import { Users } from "@/collections/Users"
+
+const filename = fileURLToPath(import.meta.url)
+const dirname = path.dirname(filename)
+
+export default buildConfig({
+  admin: {
+    user: Users.slug,
+    importMap: {
+      baseDir: path.resolve(dirname),
+    },
+  },
+  collections: [Users, Media],
+  editor: lexicalEditor(),
+  secret: process.env.PAYLOAD_SECRET || "",
+  typescript: {
+    outputFile: path.resolve(dirname, "payload-types.ts"),
+  },
+  db: sqliteAdapter({
+    // SQLite-specific arguments go here.
+    // `client.url` is required.
+    client: {
+      url: process.env.DATABASE_URL || "",
+      authToken: process.env.DATABASE_AUTH_TOKEN,
+    },
+  }),
+  sharp,
+  plugins: [
+    payloadCloudPlugin(),
+    // storage-adapter-placeholder
+  ],
+})