+++ docs/src/components/mdx/callout.tsx
@@ -6,7 +6,6 @@ import {
   LightbulbIcon,
   type LucideIcon,
 } from "lucide-react"
-import React from "react"
 
 import { cn } from "@/lib/utils"